initiate commit

+++ app.py
@@ -8,6 +8,10 @@ import os
 
 app = Flask(__name__, template_folder='templates')
 
+sm	= 0
+ldr	= 0
+t	= 0
+
 #PostgreSQL DB config----------------------------------------------
 app.config["DEBUG"] = True
 SQLALCHEMY_DATABASE_URI = "postgresql://{username}:{password}@{hostname}/{databasename}".format(
@@ -31,46 +35,52 @@ app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
 
 db = SQLAlchemy(app)
 
-value		= 0
-timestamp	= 0
-name		= 0
-
-class sensor(db.Model):
-	__tablename__ = "sensor"
-	id		= db.Column(db.Integer, primary_key=True)
-	name	= db.Column(db.String(50), default = 'sensor')
-
 class data(db.Model):
-	__tablename__ = "data"
-	id			= db.Column(db.Integer, primary_key=True)
-	value		= db.Column(db.Integer)
-	timestamp	= db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
+    __tablename__ = "data"
+    id 	= db.Column(db.Integer, primary_key=True)
+    sm 	= db.Column(db.Integer)
+    ldr = db.Column(db.Integer)
+    t	= db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
+#	def __init__(self,sm,ldr):
+#	self.sm 	= sm
+#	self.ldr	= ldr
 
 # Initialize DB manually--------------------------------------------
-def recreate_db():
-	db.drop_all()
-	db.create_all()
+engine = sa.create_engine(app.config['SQLALCHEMY_DATABASE_URI'])
+inspector = sa.inspect(engine)
+if not inspector.has_table("users"):
+	with app.app_context():
+		db.drop_all()
+		db.create_all()
+		app.logger.info('Initialized the database!')
+else:
+	app.logger.info('Database already contains the users table.')
 
 # Backend Web-------------------------------------------------------
 @app.route('/')
 def home():
 	return '<h2>Jaunty Jaugar</h2>'
 
-@app.route('/read', methods = ['POST', 'GET'])
-def read():	
+@app.route('/sensor', methods = ['POST', 'GET'])
+def sensor():
+	global sm
+	global ldr
+	global t
 	if request.method == 'POST':
-		dat	= data(
-		value = request.form.get('S1'),
-		timestamp = datetime.now(tz=timezone('Asia/Kuala_Lumpur'))
+		sm = request.form.get('sm')
+		ldr = request.form.get('ldr')
+		datadb = data(
+		sm	= request.form.get('sm'),
+		ldr	= request.form.get('ldr'),
+		t	= datetime.now(tz=timezone('Asia/Kuala_Lumpur'))
 		)
-		db.session.add(dat)
+		db.session.add(datadb)
 		db.session.commit()
-	return render_template('data.html', s1 = request.form.get('S1'))
+	return render_template('sensor.html', sm=sm, ldr=ldr)
 		
 #	else:
 #		return "<h2>ERROR</h2>"
 
 if __name__ == "__main__":
     port = os.environ.get("PORT", 5000)# Get port number of env at runtime, else use default port 5000
-    recreate_db()
     app.run(host='0.0.0.0', port=port, debug=True)
